Consolidate Share bar components in <ShareBar />
